+++ pgcrypto/fields.py
@@ -3,34 +3,29 @@ from django.db import models
 from pgcrypto import (
     DIGEST_SQL,
     HMAC_SQL,
-    INTEGER_PGP_PUB_ENCRYPT_SQL,
-    INTEGER_PGP_SYM_ENCRYPT_SQL,
+    PGP_PUB_ENCRYPT_SQL_WITH_NULLIF,
+    PGP_SYM_ENCRYPT_SQL_WITH_NULLIF,
 )
 from pgcrypto.lookups import (
-    DateTimeExactLookup,
-    DateTimeGteLookup,
-    DateTimeGtLookup,
-    DateTimeLteLookup,
-    DateTimeLtLookup,
-    DateTimeRangeLookup,
     HashLookup,
 )
 from pgcrypto.mixins import (
-    DatePGPPublicKeyFieldMixin,
-    DatePGPSymmetricKeyFieldMixin,
-    DateTimePGPPublicKeyFieldMixin,
-    DateTimePGPSymmetricKeyFieldMixin,
-    EmailPGPPublicKeyFieldMixin,
-    EmailPGPSymmetricKeyFieldMixin,
+    DecimalPGPFieldMixin,
+    get_setting,
     HashMixin,
     PGPPublicKeyFieldMixin,
     PGPSymmetricKeyFieldMixin,
+    RemoveMaxLengthValidatorMixin,
 )
 
 
 class TextDigestField(HashMixin, models.TextField):
     """Text digest field for postgres."""
     encrypt_sql = DIGEST_SQL
+
+    def get_encrypt_sql(self, connection):
+        """Get encrypt sql."""
+        return self.encrypt_sql.format(get_setting(connection, 'PGCRYPTO_KEY'))
 
 
 TextDigestField.register_lookup(HashLookup)
@@ -44,13 +39,14 @@ class TextHMACField(HashMixin, models.TextField):
 TextHMACField.register_lookup(HashLookup)
 
 
-class EmailPGPPublicKeyField(EmailPGPPublicKeyFieldMixin, models.EmailField):
+class EmailPGPPublicKeyField(RemoveMaxLengthValidatorMixin,
+                             PGPSymmetricKeyFieldMixin, models.EmailField):
     """Email PGP public key encrypted field."""
 
 
 class IntegerPGPPublicKeyField(PGPPublicKeyFieldMixin, models.IntegerField):
     """Integer PGP public key encrypted field."""
-    encrypt_sql = INTEGER_PGP_PUB_ENCRYPT_SQL
+    encrypt_sql = PGP_PUB_ENCRYPT_SQL_WITH_NULLIF
     cast_type = 'INT4'
 
 
@@ -58,37 +54,26 @@ class TextPGPPublicKeyField(PGPPublicKeyFieldMixin, models.TextField):
     """Text PGP public key encrypted field."""
 
 
-class DatePGPPublicKeyField(DatePGPPublicKeyFieldMixin, models.TextField):
+class DatePGPPublicKeyField(PGPPublicKeyFieldMixin, models.DateField):
     """Date PGP public key encrypted field for postgres."""
+    encrypt_sql = PGP_PUB_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'DATE'
 
 
-DatePGPPublicKeyField.register_lookup(DateTimeExactLookup)
-DatePGPPublicKeyField.register_lookup(DateTimeLtLookup)
-DatePGPPublicKeyField.register_lookup(DateTimeLteLookup)
-DatePGPPublicKeyField.register_lookup(DateTimeGtLookup)
-DatePGPPublicKeyField.register_lookup(DateTimeGteLookup)
-DatePGPPublicKeyField.register_lookup(DateTimeRangeLookup)
-
-
-class DateTimePGPPublicKeyField(DateTimePGPPublicKeyFieldMixin, models.TextField):
+class DateTimePGPPublicKeyField(PGPPublicKeyFieldMixin, models.DateTimeField):
     """DateTime PGP public key encrypted field for postgres."""
+    encrypt_sql = PGP_PUB_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'TIMESTAMP'
 
 
-DateTimePGPPublicKeyField.register_lookup(DateTimeExactLookup)
-DateTimePGPPublicKeyField.register_lookup(DateTimeLtLookup)
-DateTimePGPPublicKeyField.register_lookup(DateTimeLteLookup)
-DateTimePGPPublicKeyField.register_lookup(DateTimeGtLookup)
-DateTimePGPPublicKeyField.register_lookup(DateTimeGteLookup)
-DateTimePGPPublicKeyField.register_lookup(DateTimeRangeLookup)
-
-
-class EmailPGPSymmetricKeyField(EmailPGPSymmetricKeyFieldMixin, models.EmailField):
+class EmailPGPSymmetricKeyField(RemoveMaxLengthValidatorMixin,
+                                PGPSymmetricKeyFieldMixin, models.EmailField):
     """Email PGP symmetric key encrypted field."""
 
 
 class IntegerPGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.IntegerField):
     """Integer PGP symmetric key encrypted field."""
-    encrypt_sql = INTEGER_PGP_SYM_ENCRYPT_SQL
+    encrypt_sql = PGP_SYM_ENCRYPT_SQL_WITH_NULLIF
     cast_type = 'INT4'
 
 
@@ -96,25 +81,47 @@ class TextPGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.TextField):
     """Text PGP symmetric key encrypted field for postgres."""
 
 
-class DatePGPSymmetricKeyField(DatePGPSymmetricKeyFieldMixin, models.TextField):
+class DatePGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.DateField):
     """Date PGP symmetric key encrypted field for postgres."""
+    encrypt_sql = PGP_SYM_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'DATE'
 
 
-DatePGPSymmetricKeyField.register_lookup(DateTimeExactLookup)
-DatePGPSymmetricKeyField.register_lookup(DateTimeLtLookup)
-DatePGPSymmetricKeyField.register_lookup(DateTimeLteLookup)
-DatePGPSymmetricKeyField.register_lookup(DateTimeGtLookup)
-DatePGPSymmetricKeyField.register_lookup(DateTimeGteLookup)
-DatePGPSymmetricKeyField.register_lookup(DateTimeRangeLookup)
-
-
-class DateTimePGPSymmetricKeyField(DateTimePGPSymmetricKeyFieldMixin, models.TextField):
+class DateTimePGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.DateTimeField):
     """DateTime PGP symmetric key encrypted field for postgres."""
+    encrypt_sql = PGP_SYM_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'TIMESTAMP'
 
 
-DateTimePGPSymmetricKeyField.register_lookup(DateTimeExactLookup)
-DateTimePGPSymmetricKeyField.register_lookup(DateTimeLtLookup)
-DateTimePGPSymmetricKeyField.register_lookup(DateTimeLteLookup)
-DateTimePGPSymmetricKeyField.register_lookup(DateTimeGtLookup)
-DateTimePGPSymmetricKeyField.register_lookup(DateTimeGteLookup)
-DateTimePGPSymmetricKeyField.register_lookup(DateTimeRangeLookup)
+class DecimalPGPPublicKeyField(DecimalPGPFieldMixin,
+                               PGPPublicKeyFieldMixin, models.DecimalField):
+    """Decimal PGP public key encrypted field for postgres."""
+
+
+class DecimalPGPSymmetricKeyField(DecimalPGPFieldMixin,
+                                  PGPSymmetricKeyFieldMixin, models.DecimalField):
+    """Decimal PGP symmetric key encrypted field for postgres."""
+
+
+class FloatPGPPublicKeyField(PGPPublicKeyFieldMixin, models.FloatField):
+    """Float PGP public key encrypted field for postgres."""
+    encrypt_sql = PGP_PUB_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'DOUBLE PRECISION'
+
+
+class FloatPGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.FloatField):
+    """Float PGP symmetric key encrypted field for postgres."""
+    encrypt_sql = PGP_SYM_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'DOUBLE PRECISION'
+
+
+class TimePGPPublicKeyField(PGPPublicKeyFieldMixin, models.TimeField):
+    """Time PGP public key encrypted field for postgres."""
+    encrypt_sql = PGP_PUB_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'TIME'
+
+
+class TimePGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.TimeField):
+    """Float PGP symmetric key encrypted field for postgres."""
+    encrypt_sql = PGP_SYM_ENCRYPT_SQL_WITH_NULLIF
+    cast_type = 'TIME'