+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# CHANGELOG
+
+## 2.5.0
+
+* Added new DecimalFields for both public and symmetric key (#64)
+* Added new FloatFields for both public and symmetric key (#64)
+* Added new TimeFields for both public and symmetric key (#64)
+* Added support for different keys based on database (#67)
+
 ## 2.4.0
 
 * Added auto-decryption of all encrypted fields including FK tables